Moved WarRepository in new war sub-package
 Changes to be committed:
	renamed:    src/main/java/com/ardaslegends/repository/WarRepository.java -> src/main/java/com/ardaslegends/repository/war/WarRepository.java
	modified:   src/main/java/com/ardaslegends/service/war/BattleService.java
	modified:   src/main/java/com/ardaslegends/service/war/WarService.java
